managing state with Context Api and useState
