+++ src/components/CartModal.jsx
@@ -3,10 +3,7 @@ import { createPortal } from 'react-dom';
 import Cart from './Cart';
 import P from 'prop-types';
 
-const CartModal = forwardRef(function Modal(
-  { cartItems, onUpdateCartItemQuantity, title, actions },
-  ref,
-) {
+const CartModal = forwardRef(function Modal({ title, actions }, ref) {
   const dialog = useRef();
 
   useImperativeHandle(ref, () => {
@@ -20,7 +17,7 @@ const CartModal = forwardRef(function Modal(
   return createPortal(
     <dialog id="modal" ref={dialog}>
       <h2>{title}</h2>
-      <Cart items={cartItems} onUpdateItemQuantity={onUpdateCartItemQuantity} />
+      <Cart />
       <form method="dialog" id="modal-actions">
         {actions}
       </form>
@@ -32,8 +29,6 @@ const CartModal = forwardRef(function Modal(
 export default CartModal;
 
 CartModal.propTypes = {
-  cartItems: P.array,
-  onUpdateCartItemQuantity: P.func,
   title: P.string,
   actions: P.node,
 };
