+++ src/components/Header.jsx
@@ -1,11 +1,12 @@
-import { useRef } from 'react';
-import P from 'prop-types';
+import { useContext, useRef } from 'react';
 import CartModal from './CartModal.jsx';
+import { CartContext } from '../store/shopping-cart-context.jsx';
 
-export default function Header({ cart, onUpdateCartItemQuantity }) {
+export default function Header() {
   const modal = useRef();
+  const { items } = useContext(CartContext);
 
-  const cartQuantity = cart.items.length;
+  const cartQuantity = items.length;
 
   function handleOpenCartClick() {
     modal.current.open();
@@ -24,13 +25,7 @@ export default function Header({ cart, onUpdateCartItemQuantity }) {
 
   return (
     <>
-      <CartModal
-        ref={modal}
-        cartItems={cart.items}
-        onUpdateCartItemQuantity={onUpdateCartItemQuantity}
-        title="Your Cart"
-        actions={modalActions}
-      />
+      <CartModal ref={modal} title="Your Cart" actions={modalActions} />
       <header id="main-header">
         <div id="main-title">
           <img src="logo.png" alt="Elegant model" />
@@ -43,10 +38,3 @@ export default function Header({ cart, onUpdateCartItemQuantity }) {
     </>
   );
 }
-
-Header.propTypes = {
-  cart: P.shape({
-    items: P.array,
-  }),
-  onUpdateCartItemQuantity: P.func,
-};
