+++ src/components/Product.jsx
@@ -1,13 +1,10 @@
 import P from 'prop-types';
+import { useContext } from 'react';
+import { CartContext } from '../store/shopping-cart-context';
 
-export default function Product({
-  id,
-  image,
-  title,
-  price,
-  description,
-  onAddToCart,
-}) {
+export default function Product({ id, image, title, price, description }) {
+  const { AddItemToCart } = useContext(CartContext);
+
   return (
     <article className="product">
       <img src={image} alt={title} />
@@ -18,7 +15,7 @@ export default function Product({
           <p>{description}</p>
         </div>
         <p className="product-actions">
-          <button onClick={() => onAddToCart(id)}>Add to Cart</button>
+          <button onClick={() => AddItemToCart(id)}>Add to Cart</button>
         </p>
       </div>
     </article>
@@ -31,5 +28,4 @@ Product.propTypes = {
   title: P.string,
   price: P.number,
   description: P.string,
-  onAddToCart: P.func,
 };
